pyproject.toml
[project]
name = "geass"
version = "0.1.2"
readme = "README.md"
description = "Command your transcription API with Geass"
authors = [
    { name = "Tony Kabilan Okeke", email = "tonykabilanokeke@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.2",
    "pillow>=10.4.0",
    "ffmpeg-python==0.2.0",
    "httpx==0.27.0",
    "pydantic==2.6.4",
    "pydantic-settings==2.2.1",
    "rich==13.7.1",
    "tqdm==4.66.2",
    "typer==0.9.0",
    "loguru==0.7.2",
    "html2text==2024.2.26",
    "transformers==4.46.1",
    "torch==2.3.1",
    "huggingface-hub==0.23.5",
    "modal>=0.65.5",
    "logfire>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "ipython-icat>=0.2.2",
    "ipython>=8.18.1",
]
test = [
    "pytest>=8.3.3",
]

[project.scripts]
geass = "geass.cli.main:app"

[project.urls]
Homepage = "https://github.com/Kabilan108/geass"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geass"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
# exclude = [
#     "/src/geass/service",
# ]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
